Viewer now uses a status bar instead of a toolbar (#358)
* Implementing a footer for the ViewerView, gradually removing elements from the viewer toolbars

* Observing to events from Footer in the Controller instead of the View

* Encoding now is a strong enumeration

* Moved the encoding popup into the footer

* Moved the word wrapping setting into the footer

* Build fix

* Use rasterized symbol in PNG instead of SVG to support 10.15

* Moved the file position indicator into the footer

* Moved the search field into a floating search view, removed the toolbar entirely

* Cosmetics, removed the focus ring around the viewer in a sheet mode

* Correct limactl path

* Allowed content hugging in the footer, hide elements depending on the view mode, added vertical separators

* Changed wrapping icons size 16->12

+++ Source/Viewer/include/Viewer/Modes.h
@@ -1,9 +1,12 @@
-// Copyright (C) 2019 Michael Kazakov. Subject to GNU General Public License version 3.
+// Copyright (C) 2019-2024 Michael Kazakov. Subject to GNU General Public License version 3.
 #pragma once
+#include <stdint.h>
 
 namespace nc::viewer {
 
-enum class ViewMode : int { // changing this values may cause stored history corruption
+// Enumeration of the possible viewer modes.
+// NB! changing this values may cause corruption of the stored history.
+enum class ViewMode : uint8_t {
     Text = 0,
     Hex = 1,
     Preview = 2
